Extract input into BrowserInputHandler
