+++ src/Browser/Avalonia.Browser/AvaloniaView.cs
@@ -24,9 +24,7 @@ namespace Avalonia.Browser
 {
     public class AvaloniaView
     {
-        private static readonly PooledList<RawPointerPoint> s_intermediatePointsPooledList = new(ClearMode.Never);
-        private readonly BrowserTopLevelImpl _topLevelImpl;
-        private EmbeddableControlRoot _topLevel;
+        private readonly EmbeddableControlRoot _topLevel;
 
         /// <param name="divId">ID of the html element where avalonia content should be rendered.</param>
         public AvaloniaView(string divId)
@@ -53,24 +51,11 @@ namespace Avalonia.Browser
             var inputElement = hostContent.GetPropertyAsJSObject("inputElement")
                             ?? throw new InvalidOperationException("InputElement cannot be null");
 
-            _topLevelImpl = new BrowserTopLevelImpl(host, nativeControlsContainer, inputElement);
-
-            _topLevel = new EmbeddableControlRoot(_topLevelImpl);
-
-            InputHelper.InitializeBackgroundHandlers();
-
-            InputHelper.SubscribeKeyEvents(
-                host,
-                OnKeyDown,
-                OnKeyUp);
-
-            InputHelper.SubscribePointerEvents(host, OnPointerMove, OnPointerDown, OnPointerUp,
-                OnPointerCancel, OnWheel);
-
-            InputHelper.SubscribeDropEvents(host, OnDragEvent);
-            InputHelper.FocusElement(host);
+            var topLevelImpl = new BrowserTopLevelImpl(host, nativeControlsContainer, inputElement);
+            _topLevel = new EmbeddableControlRoot(topLevelImpl);
 
             _topLevel.Prepare();
+            _topLevel.GotFocus += (_, _) => InputHelper.FocusElement(host);
             _topLevel.Renderer.Start(); // TODO: use Start+StopRenderer() instead.
             _topLevel.RequestAnimationFrame(_ =>
             {
@@ -93,216 +78,5 @@ namespace Avalonia.Browser
         }
 
         internal TopLevel TopLevel => _topLevel;
-
-        private static RawPointerPoint ExtractRawPointerFromJSArgs(JSObject args)
-        {
-            var point = new RawPointerPoint
-            {
-                Position = new Point(args.GetPropertyAsDouble("offsetX"), args.GetPropertyAsDouble("offsetY")),
-                Pressure = (float)args.GetPropertyAsDouble("pressure"),
-                XTilt = (float)args.GetPropertyAsDouble("tiltX"),
-                YTilt = (float)args.GetPropertyAsDouble("tiltY"),
-                Twist = (float)args.GetPropertyAsDouble("twist")
-            };
-
-            return point;
-        }
-
-        private bool OnPointerMove(JSObject args)
-        {
-            var pointerType = args.GetPropertyAsString("pointerType");
-            var point = ExtractRawPointerFromJSArgs(args);
-            var type = pointerType switch
-            {
-                "touch" => RawPointerEventType.TouchUpdate,
-                _ => RawPointerEventType.Move
-            };
-
-            var coalescedEvents = new Lazy<IReadOnlyList<RawPointerPoint>?>(() =>
-            {
-                var points = InputHelper.GetCoalescedEvents(args);
-                s_intermediatePointsPooledList.Clear();
-                s_intermediatePointsPooledList.Capacity = points.Length - 1;
-
-                // Skip the last one, as it is already processed point.
-                for (var i = 0; i < points.Length - 1; i++)
-                {
-                    var point = points[i];
-                    s_intermediatePointsPooledList.Add(ExtractRawPointerFromJSArgs(point));
-                }
-
-                return s_intermediatePointsPooledList;
-            });
-
-            return _topLevelImpl.RawPointerEvent(type, pointerType!, point, GetModifiers(args), args.GetPropertyAsInt32("pointerId"), coalescedEvents);
-        }
-
-        private bool OnPointerDown(JSObject args)
-        {
-            var pointerType = args.GetPropertyAsString("pointerType") ?? "mouse";
-            var type = pointerType switch
-            {
-                "touch" => RawPointerEventType.TouchBegin,
-                _ => args.GetPropertyAsInt32("button") switch
-                {
-                    0 => RawPointerEventType.LeftButtonDown,
-                    1 => RawPointerEventType.MiddleButtonDown,
-                    2 => RawPointerEventType.RightButtonDown,
-                    3 => RawPointerEventType.XButton1Down,
-                    4 => RawPointerEventType.XButton2Down,
-                    5 => RawPointerEventType.XButton1Down, // should be pen eraser button,
-                    _ => RawPointerEventType.Move
-                }
-            };
-
-            var point = ExtractRawPointerFromJSArgs(args);
-            return _topLevelImpl.RawPointerEvent(type, pointerType, point, GetModifiers(args), args.GetPropertyAsInt32("pointerId"));
-        }
-
-        private bool OnPointerUp(JSObject args)
-        {
-            var pointerType = args.GetPropertyAsString("pointerType") ?? "mouse";
-            var type = pointerType switch
-            {
-                "touch" => RawPointerEventType.TouchEnd,
-                _ => args.GetPropertyAsInt32("button") switch
-                {
-                    0 => RawPointerEventType.LeftButtonUp,
-                    1 => RawPointerEventType.MiddleButtonUp,
-                    2 => RawPointerEventType.RightButtonUp,
-                    3 => RawPointerEventType.XButton1Up,
-                    4 => RawPointerEventType.XButton2Up,
-                    5 => RawPointerEventType.XButton1Up, // should be pen eraser button,
-                    _ => RawPointerEventType.Move
-                }
-            };
-
-            var point = ExtractRawPointerFromJSArgs(args);
-            return _topLevelImpl.RawPointerEvent(type, pointerType, point, GetModifiers(args), args.GetPropertyAsInt32("pointerId"));
-        }
-        
-        private bool OnPointerCancel(JSObject args)
-        {
-            var pointerType = args.GetPropertyAsString("pointerType") ?? "mouse";
-            if (pointerType == "touch")
-            {
-                var point = ExtractRawPointerFromJSArgs(args);
-                _topLevelImpl.RawPointerEvent(RawPointerEventType.TouchCancel, pointerType, point,
-                    GetModifiers(args), args.GetPropertyAsInt32("pointerId"));
-            }
-
-            return false;
-        }
-
-        private bool OnWheel(JSObject args)
-        {
-            return _topLevelImpl.RawMouseWheelEvent(new Point(args.GetPropertyAsDouble("offsetX"), args.GetPropertyAsDouble("offsetY")),
-                new Vector(-(args.GetPropertyAsDouble("deltaX") / 50), -(args.GetPropertyAsDouble("deltaY") / 50)), GetModifiers(args));
-        }
-
-        private static RawInputModifiers GetModifiers(JSObject e)
-        {
-            var modifiers = RawInputModifiers.None;
-
-            if (e.GetPropertyAsBoolean("ctrlKey"))
-                modifiers |= RawInputModifiers.Control;
-            if (e.GetPropertyAsBoolean("altKey"))
-                modifiers |= RawInputModifiers.Alt;
-            if (e.GetPropertyAsBoolean("shiftKey"))
-                modifiers |= RawInputModifiers.Shift;
-            if (e.GetPropertyAsBoolean("metaKey"))
-                modifiers |= RawInputModifiers.Meta;
-
-            var buttons = e.GetPropertyAsInt32("buttons");
-            if ((buttons & 1L) == 1)
-                modifiers |= RawInputModifiers.LeftMouseButton;
-
-            if ((buttons & 2L) == 2)
-                modifiers |= e.GetPropertyAsString("type") == "pen" ? RawInputModifiers.PenBarrelButton : RawInputModifiers.RightMouseButton;
-
-            if ((buttons & 4L) == 4)
-                modifiers |= RawInputModifiers.MiddleMouseButton;
-
-            if ((buttons & 8L) == 8)
-                modifiers |= RawInputModifiers.XButton1MouseButton;
-
-            if ((buttons & 16L) == 16)
-                modifiers |= RawInputModifiers.XButton2MouseButton;
-
-            if ((buttons & 32L) == 32)
-                modifiers |= RawInputModifiers.PenEraser;
-
-            return modifiers;
-        }
-
-        public bool OnDragEvent(JSObject args)
-        {
-            var eventType = args?.GetPropertyAsString("type") switch
-            {
-                "dragenter" => RawDragEventType.DragEnter,
-                "dragover" => RawDragEventType.DragOver,
-                "dragleave" => RawDragEventType.DragLeave,
-                "drop" => RawDragEventType.Drop,
-                _ => (RawDragEventType)(int)-1
-            };
-            var dataObject = args?.GetPropertyAsJSObject("dataTransfer");
-            if (args is null || eventType < 0 || dataObject is null)
-            {
-                return false;
-            }
-
-            // If file is dropped, we need storage js to be referenced.
-            // TODO: restructure JS files, so it's not needed.
-            _ = AvaloniaModule.ImportStorage();
-
-            var position = new Point(args.GetPropertyAsDouble("offsetX"), args.GetPropertyAsDouble("offsetY"));
-            var modifiers = GetModifiers(args);
-
-            var effectAllowedStr = dataObject.GetPropertyAsString("effectAllowed") ?? "none";
-            var effectAllowed = DragDropEffects.None;
-            if (effectAllowedStr.Contains("copy", StringComparison.OrdinalIgnoreCase))
-            {
-                effectAllowed |= DragDropEffects.Copy;
-            }
-            if (effectAllowedStr.Contains("link", StringComparison.OrdinalIgnoreCase))
-            {
-                effectAllowed |= DragDropEffects.Link;
-            }
-            if (effectAllowedStr.Contains("move", StringComparison.OrdinalIgnoreCase))
-            {
-                effectAllowed |= DragDropEffects.Move;
-            }
-            if (effectAllowedStr.Equals("all", StringComparison.OrdinalIgnoreCase))
-            {
-                effectAllowed |= DragDropEffects.Move | DragDropEffects.Copy | DragDropEffects.Link;
-            }
-            if (effectAllowed == DragDropEffects.None)
-            {
-                return false;
-            }
-
-            var dropEffect = _topLevelImpl.RawDragEvent(eventType, position, modifiers, new BrowserDataObject(dataObject), effectAllowed);
-            dataObject.SetProperty("dropEffect", dropEffect.ToString().ToLowerInvariant());
-
-            return eventType is RawDragEventType.Drop or RawDragEventType.DragOver
-                   && dropEffect != DragDropEffects.None;
-        }
-        
-        private bool OnKeyDown (string code, string key, int modifier)
-        {
-            var handled = _topLevelImpl.RawKeyboardEvent(RawKeyEventType.KeyDown, code, key, (RawInputModifiers)modifier);
-
-            if (!handled && key.Length == 1)
-            {
-                handled = _topLevelImpl.RawTextEvent(key);
-            }
-
-            return handled;
-        }
-
-        private bool OnKeyUp(string code, string key, int modifier)
-        {
-            return _topLevelImpl.RawKeyboardEvent(RawKeyEventType.KeyUp, code, key, (RawInputModifiers)modifier);
-        }
     }
 }
